Refactor: Directory structure for screens & components
